+++ src/mag.c
@@ -952,7 +952,7 @@ SEXP R_flint_mag_ops1(SEXP s_op, SEXP s_x, SEXP s_dots)
 	case 54: /*    "prod" */
 	{
 		SEXP ans = newObject("mag");
-		size_t s = (op == 52) ? 2 : 1;
+		unsigned long int s = (op == 52) ? 2 : 1;
 		mag_ptr z = (mag_ptr) flint_calloc(s, sizeof(mag_t));
 		R_flint_set(ans, z, s, (R_CFinalizer_t) &R_flint_mag_finalize);
 		switch (op) {
